+++ led.h
@@ -1,0 +1,10 @@
+#ifndef _LED_H_
+#define _LED_H_
+
+#include "mmap.h"
+
+void server_init_led(void);
+
+void server_ready_led(void);
+
+#endif
